refactored util function out for tests
