+++ pyjob/tests/test_cexec.py
@@ -1,3 +1,5 @@
+__author__ = 'Felix Simkovic'
+
 import os
 import pytest
 import sys
